Moved motioncorr metadata operations to mgui. Separated motioncorr MGUI from old magicgui.
