+++ src/Ot2Rec/magicgui.py
@@ -24,86 +24,6 @@ from magicgui import magicgui as mg
     result_widget=False,
 
     project_name={"label": "Project name *"},
-    pixel_size={"label": "Pixel size (A) *",
-                "step": 0.001},
-    output_folder={"label": "MC2 output folder"},
-    file_prefix={"label": "File prefix (if different from project name)"},
-    exec_path={"label": "Path to MC2 executable"},
-    jobs_per_gpu={"label": "Jobs per GPU",
-                  "min": 1},
-    gpu_mem_usage={"label": "GPU memory usage (if applicable)",
-                #    "widget_type": "FloatSlider",
-                   "min": 0.1,
-                   "max": 2.0},
-    use_gain={"label": "Use gain reference?"},
-    gain={"label": "Gain reference file (if applicable)",
-          "widget_type": "FileEdit",
-          "mode": "w"},
-    super_res={"label": "Super-resolution images?"},
-    discard_top={"label": "# Frames discarded FROM TOP of images",
-                 "min": 0},
-    discard_bottom={"label": "# Frames discarded FROM BOTTOM of images",
-                    "min": 0},
-    tolerance={"label": "Alignment error threshold (in pixels)",
-               "min": 0},
-    max_iter={"label": "Maximum MC2 iterations",
-              "min": 1},
-    patch_size={"widget_type": "LiteralEvalLineEdit",
-                "label": "Patch configurations (Nx, Ny, %overlap)"},
-    use_subgroups={"label": "Use subgroups in alignments"}
-)
-def get_args_mc2(
-        project_name="",
-        pixel_size=0.0,
-        output_folder=Path(Path.cwd() / "motioncor"),
-        file_prefix="",
-        exec_path=Path("/opt/lmod/modules/motioncor2/1.4.0/MotionCor2_1.4.0/MotionCor2_1.4.0_Cuda110"),
-        jobs_per_gpu=2,
-        gpu_mem_usage=1.0,
-        use_gain=False,
-        gain="",
-        super_res=False,
-        discard_top=0,
-        discard_bottom=0,
-        tolerance=0.5,
-        max_iter=10,
-        patch_size=[5,5,20],
-        use_subgroups=True,
-):
-    """
-    Function to add arguments to parser for MotionCor
-
-    ARGS:
-    project_name (str)    :: Name of current project
-    pixel_size (float)    :: Image pixel size in Angstroms
-    output_folder (str)   :: Path to folder for storing motion-corrected images (Default: ./motioncor/)
-    file_prefix (str)     :: Common prefix of raw image files (Default: project)
-    no_gpu (bool)         :: Use CPU only for motion-correction
-    jobs_per_gpu (int)    :: Number of job instance(s) per GPU
-    gpu_mem_usage (float) :: MotionCor2 GPU memory usage
-    exec_path (str)       :: Path to MotionCor2 executable
-    use_gain (bool)       :: Whether to use gain reference file
-    gain (str)            :: Path to gain reference file (leave blank if use_gain==False)
-    super_res (bool)      :: True if super-resolution images used
-    discard_top (int)     :: Number of frames discarded from top per image
-    discard_bottom (int)  :: Number of frames discarded from bottom per image
-    tolerance (float)     :: Threshold of alignment errors in pixels
-    max_iter (int)        :: Maximum number of iterations performed by MotionCor2
-    patch_size (int)      :: Size of patches used in alignment
-    use_subgroups (bool)  :: Use subgroups in alignment
-
-    OUTPUTs:
-    Namespace
-    """
-    return locals()
-
-
-@mg(
-    call_button="Create config file",
-    layout="vertical",
-    result_widget=False,
-
-    project_name={"label": "Project name *"},
     output_folder={"label": "CTFFind4 output folder",
                    "mode": "d"},
     file_prefix={"label": "File prefix (if different from project name)"},
