+++ src/Ot2Rec/params.py
@@ -71,28 +71,28 @@ def new_mc2_yaml(args):
     args (Namespace) :: Namespace generated with user inputs
     """
 
-    mc2_yaml_name = args.project_name.value + '_mc2.yaml'
+    mc2_yaml_name = args.project_name + '_mc2.yaml'
 
     mc2_yaml_dict = {
         'System': {
             'process_list': None,
-            'output_path': str(args.output_folder.value),
-            'output_prefix': args.file_prefix.value if args.file_prefix.value != "" else args.project_name.value,
-            'use_gpu': 'auto', # if not args.no_gpu.value else False,
-            'jobs_per_gpu': args.jobs_per_gpu.value,
-            'gpu_memory_usage': args.gpu_mem_usage.value,
+            'output_path': str(args.output_folder),
+            'output_prefix': args.file_prefix if args.file_prefix != "" else args.project_name,
+            'use_gpu': 'auto', # if not args.no_gpu else False,
+            'jobs_per_gpu': args.jobs_per_gpu,
+            'gpu_memory_usage': args.gpu_mem_usage,
         },
         'MC2': {
-            'MC2_path': str(args.exec_path.value),
-            'gain_reference': 'nogain' if not args.use_gain.value else str(args.gain.value),
-            'pixel_size': args.pixel_size.value,
-            'desired_pixel_size': args.pixel_size.value * 2 if args.super_res.value else args.pixel_size.value,
-            'discard_frames_top': args.discard_top.value,
-            'discard_frames_bottom': args.discard_bottom.value,
-            'tolerance': args.tolerance.value,
-            'max_iterations': args.max_iter.value,
-            'patch_size': args.patch_size.value,
-            'use_subgroups': args.use_subgroups.value,
+            'MC2_path': str(args.exec_path),
+            'gain_reference': 'nogain' if not args.use_gain else str(args.gain),
+            'pixel_size': args.pixel_size,
+            'desired_pixel_size': args.pixel_size * 2 if args.super_res else args.pixel_size,
+            'discard_frames_top': args.discard_top,
+            'discard_frames_bottom': args.discard_bottom,
+            'tolerance': args.tolerance,
+            'max_iterations': args.max_iter,
+            'patch_size': args.patch_size,
+            'use_subgroups': args.use_subgroups,
         },
     }
